feat: implement BeaconEngine.prepareForNextSlot()
